step1: refactor: generator 구현체 impl 패키지로 이동
